requirements.txt
# Core dependencies
numpy>=1.21.0
matplotlib>=3.4.0
colorama>=0.4.6
seaborn>=0.12.0
pandas>=2.0.0

# Database 
sqlalchemy>=1.4.0

# Machine Learning & Vector Analysis
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.0.0

# Testing
pytest>=6.0.0

# Mathematics Validation
sympy>=1.9  # Symbolic mathematics

# Development tools
black>=21.0.0  # Code formatting
flake8>=3.9.0  # Code linting

# Networking and API
networkx>=2.6.0  # Graph operations
arxiv>=1.4.0
fastapi>=0.68.0
uvicorn>=0.15.0

# Environment
python-dotenv>=0.19.0